+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Export the default `filter` function for wrapping
   * deps: accepts@~1.2.2
     - deps: mime-types@~2.0.7
     - deps: negotiator@0.5.0